+++ docs/src/dev-docs/changelog.rst
@@ -19,9 +19,15 @@ Added
 
 - A method to train on mixed-stress datasets by setting stresses in non-periodic
   structures to NaN.
+- Support to train target variants defined by <base>/<variant> (i.e. ``energy/PBE``).
+  Variants can be selected as a property to be predicted by an engine besides a base
+  target (i.e. ``energy``).
 
-.. Changed
-.. #######
+Changed
+#######
+
+- Raise an error (instead of warning) if energies gradients are direct targets and do
+  not have a "non_conservative" prefix
 
 .. Removed
 .. #######